feat: spatially varying latent heat release efficiency via multiple dispatch
Add a new Lscale_Cond! method dispatching on L::AbstractArray{Float64,2}
so the condensation heating efficiency can vary by longitude and latitude.
The scalar L::Float64 method is unchanged. Spectral_Physics_Interface
updated to accept either type. Includes LscaleCond_Test experiment (10-day
run with L=0.5 in tropics ±30°, L=0 elsewhere), verified precipitation
is confined to the tropical band.

+++ src/Physics/Spectral_Physics_Interface.jl
@@ -100,7 +100,7 @@ function Spectral_Physics!(
     if physics_params["do_Lscale_Cond"] && config.moisture_processes
         grid_precip .= 0.0
 
-        L = physics_params["L"]::Float64
+        L = physics_params["L"]  # Float64 or AbstractArray{Float64,2}
         Lscale_Cond!(
             vert_coord,
             atmo_data,
